api/requirements.txt
celery==4.1.0
mock==2.0.0
pymongo==3.6.1
pytest==3.5.1
python-dateutil==2.7.3
